New GDX release 7.11.19

See merge request devel/gdx!125

+++ src/gdlib/strutilx.cpp
@@ -1,8 +1,8 @@
 /*
 * GAMS - General Algebraic Modeling System GDX API
  *
- * Copyright (c) 2017-2025 GAMS Software GmbH <support@gams.com>
- * Copyright (c) 2017-2025 GAMS Development Corp. <support@gams.com>
+ * Copyright (c) 2017-2026 GAMS Software GmbH <support@gams.com>
+ * Copyright (c) 2017-2026 GAMS Development Corp. <support@gams.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -123,7 +123,7 @@ std::string IntToNiceStrW( int64_t n, int width )
    return res;
 }
 
-std::string IntToNiceStr( int N )
+std::string IntToNiceStr( int64_t N )
 {
    return IntToNiceStrW( N, 0 );
 }
