+++ src/gdlib/strutilx.hpp
@@ -1,8 +1,8 @@
 /*
 * GAMS - General Algebraic Modeling System GDX API
  *
- * Copyright (c) 2017-2025 GAMS Software GmbH <support@gams.com>
- * Copyright (c) 2017-2025 GAMS Development Corp. <support@gams.com>
+ * Copyright (c) 2017-2026 GAMS Software GmbH <support@gams.com>
+ * Copyright (c) 2017-2026 GAMS Development Corp. <support@gams.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -56,7 +56,7 @@ std::string UpperCase( std::string_view s );
 std::string LowerCase( std::string_view s );
 
 std::string IntToNiceStrW( int64_t N, int Width );
-std::string IntToNiceStr( int N );
+std::string IntToNiceStr( int64_t N );
 std::string BlankStr( unsigned int Len );
 
 // Excel column names
